feat: redo HD loading screens (should be mostly final)
Addresses #451

Fixes (R) (rights reserved) for UK and German loading screens - "Rights Reserved" is respected casually in those places, but has no legal meaning and therefore TM has to be used.

Not like this is legally going to be important anyway, it's a fanmade recreation and probably a fascimile legal warning, but I wanted it to be authentic.