ADD: Create LICENSE
Use The Unlicense.  While I am worried that contributors/the credited may be omitted by pricks who fork/incorporate this into whatever they wanna do, I'd rather not have "restrictions" on a mod (is that even a thing???) and especially not put restrictions on what is, legally, a modification on Disney-owned content and likely owned by Disney.